reorder more, add __name__ to all get_log
