+++ scripts/create_vrt.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from dolphin._log import get_log
 from dolphin.stack import VRTStack
 
-logger = get_log()
+logger = get_log(__name__)
 
 
 def get_cli_args():
