+++ src/dolphin/_background.py
@@ -5,7 +5,7 @@ from threading import enumerate as threading_enumerate
 
 from dolphin._log import get_log
 
-logger = get_log()
+logger = get_log(__name__)
 
 _DEFAULT_TIMEOUT = 1.0
