+++ src/dolphin/workflows/sequential.py
@@ -22,7 +22,9 @@ from dolphin.interferogram import VRTInterferogram
 from dolphin.phase_link import PhaseLinkRuntimeError, compress, run_mle
 from dolphin.stack import VRTStack
 
-logger = get_log()
+from ._utils import setup_output_folder
+
+logger = get_log(__name__)
 
 __all__ = ["run_evd_sequential"]
 
@@ -105,7 +107,7 @@ def run_evd_sequential(
             f" {Path(cur_vrt.file_list[-1]).name}"
         )
         # Set up the output folder with empty files to write into
-        cur_output_files = _setup_output_folder(
+        cur_output_files = setup_output_folder(
             cur_vrt, driver="GTiff", start_idx=mini_idx, strides=strides
         )
         # Save these for the final adjustment later
@@ -234,7 +236,7 @@ def run_evd_sequential(
     )
 
     logger.info(f"Running EVD on compressed files: {adjustment_vrt_stack}")
-    adjusted_comp_slc_files = _setup_output_folder(
+    adjusted_comp_slc_files = setup_output_folder(
         adjustment_vrt_stack, driver="GTiff", strides=strides
     )
 
@@ -302,72 +304,3 @@ def run_evd_sequential(
         A=tcorr_files,
         calc="numpy.nanmean(A, axis=0)",
     )
-
-
-def _setup_output_folder(
-    vrt_stack,
-    driver: str = "GTiff",
-    dtype="complex64",
-    start_idx: int = 0,
-    strides: Dict[str, int] = {"y": 1, "x": 1},
-    creation_options: Optional[List] = None,
-) -> List[Path]:
-    """Create empty output files for each band after `start_idx` in `vrt_stack`.
-
-    Also creates an empty file for the compressed SLC.
-    Used to prepare output for block processing.
-
-    Parameters
-    ----------
-    vrt_stack : VRTStack
-        object containing the current stack of SLCs
-    driver : str, optional
-        Name of GDAL driver, by default "GTiff"
-    dtype : str, optional
-        Numpy datatype of output files, by default "complex64"
-    start_idx : int, optional
-        Index of vrt_stack to begin making output files.
-        This should match the ministack index to avoid re-creating the
-        past compressed SLCs.
-    strides : Dict[str, int], optional
-        Strides to use when creating the empty files, by default {"y": 1, "x": 1}
-        Larger strides will create smaller output files, computed using
-        [dolphin.io.compute_out_shape][]
-    creation_options : list, optional
-        List of options to pass to the GDAL driver, by default None
-
-    Returns
-    -------
-    List[Path]
-        List of saved empty files.
-    """
-    output_folder = vrt_stack.outfile.parent
-
-    date_strs: List[str] = []
-    for d in vrt_stack.dates[start_idx:]:
-        if len(d) == 1:
-            # normal SLC files will have a single date
-            s = d[0].strftime(io.DEFAULT_DATETIME_FORMAT)
-        else:
-            # Compressed SLCs will have 2 dates in the name marking the start and end
-            s = io._format_date_pair(d[0], d[1])
-        date_strs.append(s)
-
-    output_files = []
-    for filename in date_strs:
-        slc_name = Path(filename).stem
-        output_path = output_folder / f"{slc_name}.slc.tif"
-
-        io.write_arr(
-            arr=None,
-            like_filename=vrt_stack.outfile,
-            output_name=output_path,
-            driver=driver,
-            nbands=1,
-            dtype=dtype,
-            strides=strides,
-            options=creation_options,
-        )
-
-        output_files.append(output_path)
-    return output_files
